+++ inquire/Cargo.toml
@@ -1,6 +1,6 @@
 [package]
 name = "inquire"
-version = "1.0.0-dev"
+version = "0.6.1"
 description = "inquire is a library for building interactive prompts on terminals"
 repository = "https://github.com/mikaelmello/inquire"
 license = "MIT"
