+++ inquire/src/prompts/confirm/action.rs
@@ -1,1 +1,4 @@
-use crate::ui::{InnerAction, Key, KeyModifiers};
+use crate::CustomTypePromptAction;
+
+/// Set of actions for a ConfirmPrompt.
+pub type ConfirmPromptAction = CustomTypePromptAction;
