+++ inquire/src/prompts/custom_type/config.rs
@@ -1,5 +1,6 @@
 use crate::CustomType;
 
+/// Configuration settings used in the execution of a CustomTypePrompt.
 pub struct CustomTypeConfig {}
 
 impl<T> From<&CustomType<'_, T>> for CustomTypeConfig {
