+++ inquire/src/prompts/dateselect/config.rs
@@ -2,8 +2,10 @@ use chrono::NaiveDate;
 
 use crate::DateSelect;
 
+/// Configuration settings used in the execution of a DateSelectPrompt.
 #[derive(Copy, Clone, Debug)]
 pub struct DateSelectConfig {
+    /// Whether to use vim-style keybindings.
     pub vim_mode: bool,
 
     /// Min date allowed to be selected.
@@ -12,6 +14,7 @@ pub struct DateSelectConfig {
     /// Max date allowed to be selected.
     pub max_date: Option<NaiveDate>,
 
+    /// Weekday to start the week on.
     pub week_start: chrono::Weekday,
 }
