+++ inquire/src/prompts/editor/config.rs
@@ -1,9 +1,12 @@
 use crate::Editor;
 use std::ffi::OsStr;
 
+/// Configuration settings used in the execution of an EditorPrompt.
 #[derive(Copy, Clone, Debug)]
 pub struct EditorConfig<'a> {
+    /// The command to use to open the editor.
     pub editor_command: &'a OsStr,
+    /// The arguments to pass to the editor command.
     pub editor_command_args: &'a [&'a OsStr],
 }
