+++ inquire/src/prompts/mod.rs
@@ -1,3 +1,4 @@
+mod action;
 mod confirm;
 mod custom_type;
 #[cfg(feature = "date")]
@@ -6,17 +7,18 @@ mod dateselect;
 mod editor;
 mod multiselect;
 mod password;
-pub mod prompt;
+mod prompt;
 mod select;
 mod text;
 
-pub use confirm::Confirm;
-pub use custom_type::CustomType;
+pub use action::*;
+pub use confirm::*;
+pub use custom_type::*;
 #[cfg(feature = "date")]
-pub use dateselect::DateSelect;
+pub use dateselect::*;
 #[cfg(feature = "editor")]
-pub use editor::Editor;
-pub use multiselect::MultiSelect;
-pub use password::{Password, PasswordDisplayMode};
-pub use select::Select;
-pub use text::Text;
+pub use editor::*;
+pub use multiselect::*;
+pub use password::*;
+pub use select::*;
+pub use text::*;
