+++ inquire/src/prompts/multiselect/config.rs
@@ -1,9 +1,13 @@
 use crate::MultiSelect;
 
+/// Configuration settings used in the execution of a MultiSelectPrompt.
 #[derive(Copy, Clone, Debug)]
 pub struct MultiSelectConfig {
+    /// Whether to use vim-style keybindings.
     pub vim_mode: bool,
+    /// Page size of the list of options.
     pub page_size: usize,
+    /// Whether to keep the filter text when an option is selected.
     pub keep_filter: bool,
 }
