+++ inquire/src/prompts/multiselect/mod.rs
@@ -1,9 +1,11 @@
-pub mod action;
-pub mod config;
+mod action;
+mod config;
 mod prompt;
 #[cfg(test)]
 #[cfg(feature = "crossterm")]
 mod test;
+
+pub use action::*;
 
 use std::fmt::Display;
 
@@ -12,7 +14,7 @@ use crate::{
     error::{InquireError, InquireResult},
     formatter::MultiOptionFormatter,
     list_option::ListOption,
-    prompt::Prompt,
+    prompts::prompt::Prompt,
     terminal::get_default_terminal,
     type_aliases::Filter,
     ui::{Backend, MultiSelectBackend, RenderConfig},
