+++ inquire/src/prompts/multiselect/prompt.rs
@@ -3,9 +3,9 @@ use std::{collections::BTreeSet, fmt::Display};
 use crate::{
     error::InquireResult,
     formatter::MultiOptionFormatter,
-    input::{Input, InputHandleResult},
+    input::{Input, InputActionResult},
     list_option::ListOption,
-    prompt::{HandleResult, Prompt},
+    prompts::prompt::{ActionResult, Prompt},
     type_aliases::Filter,
     ui::MultiSelectBackend,
     utils::paginate,
@@ -88,7 +88,7 @@ where
             .collect()
     }
 
-    fn move_cursor_up(&mut self, qty: usize, wrap: bool) -> HandleResult {
+    fn move_cursor_up(&mut self, qty: usize, wrap: bool) -> ActionResult {
         let new_position = if wrap {
             let after_wrap = qty.saturating_sub(self.cursor_index);
             self.cursor_index
@@ -101,7 +101,7 @@ where
         self.update_cursor_position(new_position)
     }
 
-    fn move_cursor_down(&mut self, qty: usize, wrap: bool) -> HandleResult {
+    fn move_cursor_down(&mut self, qty: usize, wrap: bool) -> ActionResult {
         let mut new_position = self.cursor_index.saturating_add(qty);
 
         if new_position >= self.filtered_options.len() {
@@ -117,19 +117,19 @@ where
         self.update_cursor_position(new_position)
     }
 
-    fn update_cursor_position(&mut self, new_position: usize) -> HandleResult {
+    fn update_cursor_position(&mut self, new_position: usize) -> ActionResult {
         if new_position != self.cursor_index {
             self.cursor_index = new_position;
-            HandleResult::NeedsRedraw
+            ActionResult::NeedsRedraw
         } else {
-            HandleResult::Clean
+            ActionResult::Clean
         }
     }
 
-    fn toggle_cursor_selection(&mut self) -> HandleResult {
+    fn toggle_cursor_selection(&mut self) -> ActionResult {
         let idx = match self.filtered_options.get(self.cursor_index) {
             Some(val) => val,
-            None => return HandleResult::Clean,
+            None => return ActionResult::Clean,
         };
 
         if self.checked.contains(idx) {
@@ -142,7 +142,7 @@ where
             self.input.clear();
         }
 
-        HandleResult::NeedsRedraw
+        ActionResult::NeedsRedraw
     }
 
     fn validate_current_answer(&self) -> InquireResult<Validation> {
@@ -213,7 +213,7 @@ where
         Ok(answer)
     }
 
-    fn handle(&mut self, action: MultiSelectPromptAction) -> InquireResult<HandleResult> {
+    fn handle(&mut self, action: MultiSelectPromptAction) -> InquireResult<ActionResult> {
         let result = match action {
             MultiSelectPromptAction::MoveUp => self.move_cursor_up(1, true),
             MultiSelectPromptAction::MoveDown => self.move_cursor_down(1, true),
@@ -234,7 +234,7 @@ where
                     self.input.clear();
                 }
 
-                HandleResult::NeedsRedraw
+                ActionResult::NeedsRedraw
             }
             MultiSelectPromptAction::ClearSelections => {
                 self.checked.clear();
@@ -243,12 +243,12 @@ where
                     self.input.clear();
                 }
 
-                HandleResult::NeedsRedraw
+                ActionResult::NeedsRedraw
             }
             MultiSelectPromptAction::FilterInput(input_action) => {
                 let result = self.input.handle(input_action);
 
-                if let InputHandleResult::ContentChanged = result {
+                if let InputActionResult::ContentChanged = result {
                     let options = self.filter_options();
                     self.filtered_options = options;
                     if self.filtered_options.len() <= self.cursor_index {
