+++ inquire/src/prompts/password/config.rs
@@ -1,8 +1,11 @@
 use crate::{Password, PasswordDisplayMode};
 
+/// Configuration settings used in the execution of a PasswordPrompt.
 #[derive(Copy, Clone, Debug)]
 pub struct PasswordConfig {
+    /// Whether to allow the user to toggle the display mode of the password.
     pub enable_display_toggle: bool,
+    /// The initial display mode of the password.
     pub display_mode: PasswordDisplayMode,
 }
