+++ inquire/src/prompts/select/config.rs
@@ -1,8 +1,11 @@
 use crate::Select;
 
+/// Configuration settings used in the execution of a SelectPrompt.
 #[derive(Copy, Clone, Debug)]
 pub struct SelectConfig {
+    /// Whether to use vim-style keybindings.
     pub vim_mode: bool,
+    /// Page size of the list of options.
     pub page_size: usize,
 }
