+++ inquire/src/prompts/text/mod.rs
@@ -1,16 +1,18 @@
-pub mod action;
-pub mod config;
+mod action;
+mod config;
 mod prompt;
 #[cfg(test)]
 #[cfg(feature = "crossterm")]
 mod test;
+
+pub use action::*;
 
 use crate::{
     autocompletion::Autocomplete,
     config::get_configuration,
     error::{InquireError, InquireResult},
     formatter::{StringFormatter, DEFAULT_STRING_FORMATTER},
-    prompt::Prompt,
+    prompts::prompt::Prompt,
     terminal::get_default_terminal,
     ui::{Backend, RenderConfig, TextBackend},
     validator::StringValidator,
