+++ inquire/src/ui/mod.rs
@@ -1,6 +1,5 @@
 //! UI-related definitions for rendered content.
 
-mod action;
 mod backend;
 mod color;
 mod input_reader;
@@ -8,7 +7,6 @@ mod key;
 mod render_config;
 mod style;
 
-pub(crate) use action::*;
 pub(crate) use backend::*;
 pub(crate) use input_reader::*;
 pub(crate) use key::*;
